college_log/logs/api/filters.py
import django_filters
from logs.models import Issue


class IssueFilter(django_filters.FilterSet):
    created_after = django_filters.DateTimeFilter(field_name='created_at', lookup_expr='gte')
    created_before = django_filters.DateTimeFilter(field_name='created_at', lookup_expr='lte')

    class Meta:
        model = Issue
        fields = ['status', 'priority', 'device_type']
