college_log/requirements.txt
Django==5.1.7
django-allauth==65.11.0
django-jazzmin==3.0.1
django-simple-history==3.11.0
djangorestframework==3.15.2
drf-spectacular==0.29.0
django-filter==25.1

# Django core dependencies
asgiref==3.9.1
sqlparse==0.5.3

# allauth dependencies
requests==2.32.4
requests-oauthlib==2.0.0
python3-openid==3.2.0
oauthlib==3.3.1
